+++ examples/todo-with-better-auth/libs/transactional/src/components/heading.tsx
@@ -1,5 +1,5 @@
-import { Heading as ReactEmailHeading } from '@react-email/components';
 import * as React from 'react';
+import { Heading as ReactEmailHeading } from 'react-email';
 
 import { theme } from '../constants/theme.js';