+++ packages/project-builder-web/src/routes/data/enums/-components/enums-sidebar-list.tsx
@@ -26,7 +26,7 @@ export function EnumsSidebarList({
   className,
 }: EnumsSidebarListProps): React.JSX.Element {
   const {
-    definition: { enums = [] },
+    definition: { enums },
   } = useProjectDefinition();
 
   const [filterQuery, setFilterQuery] = useState('');