+++ packages/code-morph/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @baseplate-dev/code-morph
 
+## 0.6.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/utils@0.6.12
+
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/utils@0.6.11
+
 ## 0.6.10
 
 ### Patch Changes