+++ packages/project-builder-test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @baseplate-dev/project-builder-test
 
+## 0.6.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/project-builder-dev@0.6.12
+
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/project-builder-dev@0.6.11
+
 ## 0.6.10
 
 ### Patch Changes